backend/requirements.txt
fastapi
uvicorn[standard]
websockets
yt-dlp
openai-whisper
yfinance
anthropic
sqlite3
pydantic
python-multipart
httpx
imageio-ffmpeg
python-dotenv
